DetailImageContentを別ファイルに分割
