+++ iOSEngineerCodeCheck/Detail/DetailPresenter.swift
@@ -1,18 +1,6 @@
 import Combine
 import Foundation
 
-import class UIKit.UIImage
-
-enum DetailImageContent {
-    enum Message {
-        case loading
-        case notFound
-    }
-
-    case image(UIImage)
-    case message(Message)
-}
-
 @MainActor
 protocol ImageCacheForPresenter {
     var statePublisher: AnyPublisher<ImageCacheState, Never> { get }
